refactor(speech-to-text): update speech-to-text resource and clean up model handling
