+++ nodes/Fal/resources/speech-to-text/execute/whisper.ts
@@ -1,49 +1,48 @@
-import { IDataObject, IExecuteFunctions } from "n8n-workflow";
-import { getModelConfig, SPEECH_TO_TEXT_MODEL_IDS } from "../model";
-import { QueueSubmitResponseInterface } from "../../../interfaces";
-import { pollQueueResult } from "../../../utils/poll-queue.util";
+import type { IDataObject, IExecuteFunctions } from 'n8n-workflow';
+import { getModelConfig, SPEECH_TO_TEXT_MODEL_IDS } from '../models';
+import { QueueSubmitResponseInterface } from '../../../interfaces';
+import { pollQueueResult } from '../../../utils/poll-queue.util';
 
-
+/**
+ * Execute Whisper speech-to-text transcription
+ *
+ * Whisper is OpenAI's model for speech transcription and translation.
+ */
 export async function executeWhisper(
-    context: IExecuteFunctions, itemIndex: number
+	context: IExecuteFunctions,
+	itemIndex: number,
 ): Promise<IDataObject> {
+	// Get model configuration
+	const modelConfig = getModelConfig(SPEECH_TO_TEXT_MODEL_IDS.WHISPER);
 
-    // Get model configuration
-    const modelConfig = getModelConfig(SPEECH_TO_TEXT_MODEL_IDS.WHISPER);
+	// Get required parameters
+	const audioUrl = context.getNodeParameter('audioUrl', itemIndex) as string;
+	const language = context.getNodeParameter('language', itemIndex) as string;
+	const task = context.getNodeParameter('task', itemIndex) as string;
 
-    // Get required parameters (only required audioFileUrl)
-    const audioFileUrl = context.getNodeParameter('audioFileUrl', itemIndex) as string;
-    const language = context.getNodeParameter('language', itemIndex) as string;
-    const task = context.getNodeParameter('task', itemIndex) as string;
-    const additionalOptions = context.getNodeParameter(
-        'additionalOptions',
-        itemIndex,
-        {},
-    ) as IDataObject;
+	const body: IDataObject = {
+		audio_url: audioUrl,
+		language,
+		task,
+	};
 
-    const body: IDataObject = { //chunk_level,max_segment_len,merge_chunks,version (optional)
-        audio_file_url: audioFileUrl,
-        language,
-        task,
-    };
+	// Submit to queue
+	const submitResponse = (await context.helpers.httpRequestWithAuthentication.call(
+		context,
+		'falApi',
+		{
+			method: 'POST',
+			url: `https://queue.fal.run/${modelConfig.modelId}`,
+			body,
+			json: true,
+		},
+	)) as QueueSubmitResponseInterface;
 
-    // Submit to queue - fal.ai API 
-    const submitResponse = await context.helpers.httpRequestWithAuthentication.call(
-        context,
-        'falApi',
-        {
-            method: 'POST',
-            url: `https://queue.fal.run/${modelConfig.modelId}`,
-            body,
-            json: true,
-        },
-    ) as QueueSubmitResponseInterface;
-
-    // Poll for result using the URLs provided by the API
-    return await pollQueueResult(
-        context,
-        submitResponse.status_url,
-        submitResponse.response_url,
-        itemIndex,
-    );
+	// Poll for result using the URLs provided by the API
+	return await pollQueueResult(
+		context,
+		submitResponse.status_url,
+		submitResponse.response_url,
+		itemIndex,
+	);
 }
