+++ nodes/Fal/resources/speech-to-text/execute/wizper.ts
@@ -1,43 +1,44 @@
-import { IDataObject, IExecuteFunctions } from "n8n-workflow";
-import { getModelConfig, SPEECH_TO_TEXT_MODEL_IDS} from "../model";
-import { QueueSubmitResponseInterface } from "../../../interfaces";
-import { pollQueueResult } from "../../../utils/poll-queue.util";
+import type { IDataObject, IExecuteFunctions } from 'n8n-workflow';
+import { getModelConfig, SPEECH_TO_TEXT_MODEL_IDS } from '../models';
+import { QueueSubmitResponseInterface } from '../../../interfaces';
+import { pollQueueResult } from '../../../utils/poll-queue.util';
 
+/**
+ * Execute Wizper speech-to-text transcription
+ *
+ * Wizper is Whisper v3 Large optimized by fal.ai - same accuracy, double the performance.
+ */
 export async function executeWizper(
-    context: IExecuteFunctions, itemIndex: number
+	context: IExecuteFunctions,
+	itemIndex: number,
 ): Promise<IDataObject> {
+	// Get model configuration
+	const modelConfig = getModelConfig(SPEECH_TO_TEXT_MODEL_IDS.WIZPER);
 
-    const modelConfig = getModelConfig(SPEECH_TO_TEXT_MODEL_IDS.WIZPER);
+	// Get required parameters
+	const audioUrl = context.getNodeParameter('audioUrl', itemIndex) as string;
 
-    // Get required parameters (only required audioFileUrl)
-    const audioFileUrl = context.getNodeParameter('audioFileUrl', itemIndex) as string;
-    const additionalOptions = context.getNodeParameter(
-        'additionalOptions',
-        itemIndex,
-        {},
-    ) as IDataObject;
+	const body: IDataObject = {
+		audio_url: audioUrl,
+	};
 
-    const body: IDataObject = { //chunk_level,max_segment_len,merge_chunks,version (optional)
-        audio_file_url: audioFileUrl,
-    };
+	// Submit to queue
+	const submitResponse = (await context.helpers.httpRequestWithAuthentication.call(
+		context,
+		'falApi',
+		{
+			method: 'POST',
+			url: `https://queue.fal.run/${modelConfig.modelId}`,
+			body,
+			json: true,
+		},
+	)) as QueueSubmitResponseInterface;
 
-    // Submit to queue - fal.ai API 
-    const submitResponse = await context.helpers.httpRequestWithAuthentication.call(
-        context,
-        'falApi',
-        {
-            method: 'POST',
-            url: `https://queue.fal.run/${modelConfig.modelId}`,
-            body,
-            json: true,
-        },
-    ) as QueueSubmitResponseInterface;
-
-    // Poll for result using the URLs provided by the API
-    return await pollQueueResult(
-        context,
-        submitResponse.status_url,
-        submitResponse.response_url,
-        itemIndex,
-    );
+	// Poll for result using the URLs provided by the API
+	return await pollQueueResult(
+		context,
+		submitResponse.status_url,
+		submitResponse.response_url,
+		itemIndex,
+	);
 }
